rename store -> kvstore
