+++ cluster/kvstore/etcd.go
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package store
+package kvstore
 
 import (
 	"github.com/m3db/m3/src/cluster/client"
